Created config package and moved in basic auth, swagger config and secuity config. renaned staffs to staff in LegacyGatewayEndpoint. Added info to yamls (app and app-cloud).
