+++ src/main/java/uk/gov/ons/fwmt/gateway/config/WebSecurityConfig.java
@@ -1,11 +1,11 @@
-package uk.gov.ons.fwmt.gateway;
+package uk.gov.ons.fwmt.gateway.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-
+import uk.gov.ons.fwmt.gateway.config.BasicAuthenticationPoint;
 
 @Configuration
 @EnableWebSecurity
