+++ src/main/resources/application.yaml
@@ -1,3 +1,11 @@
+info:
+ name: ${project.artifactId}
+ version: ${project.version}
+ origin: ${scm.url}
+ commit: ${buildNumber}
+ branch: ${scmBranch}
+ built: ${timestamp}
+
 server:
   port: 9091
